refactor: split database parsing into separate files per engine
